R2 VM Core: Refactoring middlewares.
